+++ bubbleSort.h
@@ -2,5 +2,5 @@
 #include <iostream>
 using namespace std;
 
-void bubble_sort_comp(int *a ,int n, int &count_comp);
-void bubble_sort_time(int *a ,int n,long double &time);
+void bubble_sort_comp(int *a ,int n, long long &count_comp);
+void bubble_sort_time(int *a ,int n, long double &time);
