+++ shakerSort.cpp
@@ -4,7 +4,7 @@
  https://www.youtube.com/watch?v=FgExRtl66Lc
  */
 
-void shaker_sort_comp(int *a, int n, int &count_comp)
+void shaker_sort_comp(int *a, int n, long long &count_comp)
 {
     int left = 0;
     int right, mark;
